+++ src/api/client/server.rs
@@ -69,6 +69,7 @@ fn ws_index(
                         member_id,
                         Box::new(room),
                         state.config.idle_timeout,
+                        state.config.ping_interval,
                     ),
                     &request,
                     payload,