+++ proto/client-api/CHANGELOG.md
@@ -6,8 +6,16 @@ All user visible changes to this project will be documented in this file. This p
 
 
 
-## TBD [0.1.1] · 2019-??-??
-[0.1.1]: /../../tree/medea-client-api-proto-0.1.1/proto/client-api
+## TBD [0.2.0] · 2019-??-??
+[0.2.0]: /../../tree/medea-client-api-proto-0.2.0/proto/client-api
+
+### BC Breaks
+
+- RPC messages ([#75](/../../pull/75)):
+    - Server messages:
+        - `Pong` is now `Ping`.
+    - Client messages:
+        - `Ping` is now `Pong`.
 
 ### Added
 
@@ -15,7 +23,9 @@ All user visible changes to this project will be documented in this file. This p
 - `Incrementable` trait ([#28]);
 - `CloseReason` and `CloseDescription` types ([#58](/../../pull/58));
 - `AddPeerConnectionMetrics` client command with `IceConnectionState` metric ([#71](/../../pull/71));
+- `RpcSettings` server message ([#75](/../../pull/75));
 - `is_relay` into `Event::PeerCreated` ([#79](/../../pull/79)).
+>>>>>>> master
 
 [#28]: /../../pull/28